requirements.txt
streamlit
mftool
langchain
langchain-pinecone 
pinecone-notebooks
langchain-groq
langchain-mongodb
langchain-community


